# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopmee API Documentation",
        "description": "API endpoints for Shopmee e-commerce application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Shopmee API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Public Products"
                ],
                "summary": "Xem danh sách sản phẩm CÔNG KHAI",
                "description": "Lấy danh sách sản phẩm công khai hỗ trợ tìm kiếm, lọc theo giá, lọc theo danh mục, sắp xếp và phân trang.",
                "operationId": "getPublicProducts",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm theo tên sản phẩm",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price_min",
                        "in": "query",
                        "description": "Giá tối thiểu",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "price_max",
                        "in": "query",
                        "description": "Giá tối đa",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Trường sắp xếp",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "price"
                            ]
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Hướng sắp xếp",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Số lượng sản phẩm mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy danh sách sản phẩm thành công."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Đăng sản phẩm mới và tạo các biến thể",
                "description": "API dành cho Seller để đăng sản phẩm mới kèm các biến thể và ảnh tùy chọn.",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name",
                                    "description",
                                    "variants"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Áo thun Polo Classic"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mô tả chi tiết Áo thun Polo Classic"
                                    },
                                    "variants": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "sku",
                                                "variant_name",
                                                "price",
                                                "stock_quantity"
                                            ],
                                            "properties": {
                                                "sku": {
                                                    "type": "string",
                                                    "example": "POLO-W-M"
                                                },
                                                "variant_name": {
                                                    "type": "string",
                                                    "example": "White - M"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 199000
                                                },
                                                "stock_quantity": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "images": {
                                        "description": "Mảng các file ảnh upload",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo sản phẩm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng sản phẩm và tạo các biến thể thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Áo thun Polo Classic"
                                                },
                                                "variants_count": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ hoặc lỗi nghiệp vụ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bạn chưa đăng ký cửa hàng."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/products/{id}": {
            "get": {
                "tags": [
                    "Public Products"
                ],
                "summary": "Xem chi tiết một sản phẩm theo ID",
                "description": "Lấy thông tin chi tiết của sản phẩm kèm các biến thể và danh sách hình ảnh.",
                "operationId": "getPublicProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy chi tiết sản phẩm thành công."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sản phẩm không tồn tại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sản phẩm không tồn tại."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/search": {
            "get": {
                "tags": [
                    "Public Products"
                ],
                "summary": "Tìm kiếm sản phẩm nâng cao",
                "description": "Tìm kiếm sản phẩm theo tên, mô tả, biến thể, danh mục hoặc cửa hàng.",
                "operationId": "searchPublicProducts",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Số lượng sản phẩm mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tìm kiếm sản phẩm thành công."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}